+++ Sources/Sentry/Public/SentryOptions.h
@@ -328,6 +328,19 @@ NS_SWIFT_NAME(Options)
  */
 @property (nonatomic, assign) BOOL sendClientReports;
 
+/**
+ * When enabled, the SDK tracks when the application stops responding for a specific amount of
+ * time defined by the `appHangsTimeoutInterval` option.
+ */
+@property (nonatomic, assign) BOOL enableAppHangTracking;
+
+/**
+ * The minimum amount of time an app shoud be unresponsive to be classified as an App Hanging.
+ * The actual amount may be a little longer.
+ * Avoid using values lower than 100ms, which may cause a lot of app hangs events being transmitted.
+ */
+@property (nonatomic, assign) NSTimeInterval appHangTimeoutInterval;
+
 @end
 
 NS_ASSUME_NONNULL_END