+++ Sources/Sentry/SentryClient.m
@@ -477,6 +477,7 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
             || (nil != event.exceptions && [event.exceptions count] > 0);
 
         BOOL threadsNotAttached = !(nil != event.threads && event.threads.count > 0);
+
         if (!isCrashEvent && shouldAttachStacktrace && threadsNotAttached) {
             event.threads = [self.threadInspector getCurrentThreads];
         }