+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Features
 
+- Add App Hangs tracking (#1861) 
 - Replace tracestate header with baggage (#1867)
 - Implement description for SentryBreadcrumb (#1880)